bitsの追加およびWarningへの対応。
1.bits を追加。単体テストも追加、OK。
  本来inline関数にすべきかとも思うが、
  型に合わせてたくさん必要になってしまうので、
  敢えてマクロ実装。

2.Map_KeyAt()がヘッダに記載されていなかった。追加。

3.Arduino開発時-Wallで現れたWarningへの対応。

全般単体テスト再実施済み。ただしcovは省略。

+++ src/Assertions.c
@@ -105,7 +105,7 @@ void Assertions_AssertWith(
 			Assertions_Item item;
 			item.Line = line;
 			// ターミネータを含めたサイズをコピー
-			int32_t fnl = strlen(fileName) + 1;
+			size_t fnl = strlen(fileName) + 1;
 			if (fnl > sizeof item.FileName)
 			{
 				fnl = sizeof item.FileName;
